requirements.txt
pandas==2.2.3
numpy==2.2.3
openpyxl==3.1.5
PyYAML==6.0.2
